+++ src/main/java/com/bawnorton/neruina/blacklist/Blacklist.java
@@ -1,0 +1,77 @@
+package com.bawnorton.neruina.blacklist;
+
+import com.bawnorton.neruina.util.ErroredType;
+import com.google.gson.FieldNamingPolicy;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+import net.minecraft.util.Identifier;
+import java.io.IOException;
+import java.util.List;
+
+public final class Blacklist {
+    private static final Gson GSON = new GsonBuilder()
+        .setPrettyPrinting()
+        .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
+        .registerTypeAdapter(Identifier.class, new TypeAdapter<>() {
+            @Override
+            public void write(JsonWriter out, Object value) throws IOException {
+                if (value instanceof Identifier id) {
+                    out.value(id.toString());
+                } else if (value == null) {
+                    out.nullValue();
+                }
+            }
+
+            @Override
+            public Object read(JsonReader in) throws IOException {
+                String value = in.nextString();
+                Identifier identifier = Identifier.tryParse(value);
+                if (identifier == null) {
+                    throw new IOException("Invalid Identifier: " + value);
+                }
+                return identifier;
+            }
+        })
+        .create();
+
+    private List<String> namespaces;
+    private List<Identifier> entities;
+    private List<Identifier> blocks;
+    private List<Identifier> blockEntities;
+    private List<Identifier> items;
+
+    public Blacklist(List<String> namespaces, List<Identifier> entities, List<Identifier> blocks, List<Identifier> blockEntities, List<Identifier> items) {
+        this.namespaces = namespaces;
+        this.entities = entities;
+        this.blocks = blocks;
+        this.blockEntities = blockEntities;
+        this.items = items;
+    }
+
+    public static Blacklist fromJson(JsonReader reader) {
+        return GSON.fromJson(reader, Blacklist.class);
+    }
+
+    public boolean isBlacklisted(ErroredType type, Identifier id) {
+        if (id == null) {
+            return false;
+        }
+        if (namespaces != null) {
+            for (String namespace : namespaces) {
+                if (id.getNamespace().equals(namespace)) {
+                    return true;
+                }
+            }
+        }
+        return switch (type) {
+            case ENTITY -> entities != null && entities.contains(id);
+            case BLOCK_STATE -> blocks != null && blocks.contains(id);
+            case BLOCK_ENTITY -> blockEntities != null && blockEntities.contains(id);
+            case ITEM_STACK -> items != null && items.contains(id);
+            default -> false;
+        };
+    }
+}