+++ src/main/java/com/bawnorton/neruina/exception/TickingException.java
@@ -1,5 +1,7 @@
 package com.bawnorton.neruina.exception;
 
+import net.minecraft.util.Identifier;
+
 public class TickingException extends RuntimeException {
     public TickingException(String message, Throwable cause) {
         super(message, cause);
@@ -8,6 +10,16 @@ public class TickingException extends RuntimeException {
     public static TickingException notHandled(String configOption, Throwable cause) {
         return new TickingException(
                 "Ticking exception not handled as \"%s\" is set to \"false\"".formatted(configOption),
+                cause
+        );
+    }
+
+    public static TickingException blacklisted(Identifier owningBlacklist, Identifier tickingId, Throwable cause) {
+        return new TickingException(
+                "Ticking exception not handled as handling for \"%s\" is blacklisted by \"%s\"".formatted(
+                        tickingId,
+                        owningBlacklist
+                ),
                 cause
         );
     }