+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.3.0
+
+- Added datadriven blackist to allow mods or modpacks to opt-out specific blocks, items, entities, etc from handling or entire namespaces
+- Migrated messaging system to use permission level instead of discrete options (default: 0)
+- Put all Neruina commands under a required permission level controlled by the config (default: 2)
+
 # 2.2.11
 
 - Cache entity UUID to prevent world interaction when re-serialising ticking entries